Basics are laid out, need to handle async better.
The basics of the website are in place with regards to the front end. The back end is held up by my lackluster  NodeJS skills when it comes to handling promises. I learned a lot today and, while the code is an absolute mess right now, I'm feeling good about cleaning it up tomorrow.