Created a new model without the pre middleware for getPrices
When querying the database with a .find({}) in order to start getting the prices for each tradable, memory would run out. After some investigation I found the problem to be the deepPopulate. Tentatively, this is the only solution I can think of as I could not see a way to conditionally include/exclude the middleware.